ESS IDEA DB Concurrency Handled.... Simultaneously updating same deals from 2 different clients will create 2 version on first come first served basis... Execution in a DB Atomic Transaction by Acquiring Read Locks on the Rows
